feat(shared): added colors tag and current date in utils
