+++ src/shared/ui/project-card/project-card.module.scss
@@ -15,7 +15,7 @@
 	}
 
 	@media screen and (max-width: 779px) {
-		padding: 24px;
+		padding: 16px;
 		width: 280px;
 		min-height: 145px;
 	}
