+++ src/shared/utils/index.ts
@@ -1,1 +1,3 @@
-export { getClassNameforTag } from './get-tag-color/get-tag-color';
+export { getColorTag } from './get-tag-color/get-tag-color';
+export { getStartDate } from './get-current-date/get-current-date';
+export { getEndDate } from './get-current-date/get-current-date';
